Step7-9: split file
